# docs/swagger.yaml
definitions:
  helper.ErrorResponse:
    properties:
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  schema.LoginUserResponse:
    properties:
      code:
        type: integer
      jwt_token:
        type: string
      message:
        type: string
    type: object
  schema.LoginUserWithUsernameRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  schema.LoginWithGmailRequest:
    properties:
      gmail:
        type: string
      password:
        type: string
    required:
    - gmail
    - password
    type: object
  schema.RegisterUserRequest:
    properties:
      firstname:
        type: string
      gmail:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - firstname
    - gmail
    - lastname
    - password
    - username
    type: object
  schema.RegisterUserResponse:
    properties:
      firstname:
        type: string
      gmail:
        type: string
      id:
        type: string
      jwt_token:
        type: string
      lastname:
        type: string
      username:
        type: string
    type: object
  schema.WithdrawNativeRequest:
    properties:
      amount:
        type: string
      asset_owner_address:
        type: string
      calldata:
        type: string
      chain_id:
        type: integer
      destination_address:
        type: string
      gas_limit_unit:
        type: integer
      gas_price_wei:
        type: string
      token_address:
        type: string
      token_symbol:
        type: string
    required:
    - amount
    - asset_owner_address
    - chain_id
    - destination_address
    - gas_limit_unit
    - gas_price_wei
    - token_address
    - token_symbol
    type: object
  schema.WithdrawNativeResponse:
    properties:
      code:
        type: integer
      transaction_status:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/auth/login/gmail:
    post:
      consumes:
      - application/json
      description: Authenticate user using gmail and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.LoginWithGmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully logged in
          schema:
            $ref: '#/definitions/schema.LoginUserResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Login user with gmail
      tags:
      - auth
  /v1/auth/login/username:
    post:
      consumes:
      - application/json
      description: Authenticate user using username and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.LoginUserWithUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully logged in
          schema:
            $ref: '#/definitions/schema.LoginUserResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Login user with username
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, firstname, lastname, and gmail
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully registered
          schema:
            $ref: '#/definitions/schema.RegisterUserResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /v1/transaction/withdraw:
    post:
      consumes:
      - application/json
      description: Handles asset withdrawal requests for native tokens.
      parameters:
      - description: WithdrawNativeRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.WithdrawNativeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful withdrawal
          schema:
            $ref: '#/definitions/schema.WithdrawNativeResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: User or wallet not found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Withdraw asset
      tags:
      - transaction
swagger: "2.0"